Floats, not ints
I can see that coming up in the future, e.g. with decimal temperature values

+++ testscript.js
@@ -16,8 +16,8 @@ function othertestfunction() {
 function add_these_two_numbers() {
 	var numform = document.getElementById("testform");
 	var numlist = numform.querySelectorAll("input[type=number]");
-	var firstnum = parseInt(numlist[0].value);
-	var secondnum = parseInt(numlist[1].value);
+	var firstnum = parseFloat(numlist[0].value);
+	var secondnum = parseFloat(numlist[1].value);
 	var sum_of_nums = firstnum + secondnum;
 	document.getElementById("resultdump").innerHTML = sum_of_nums;
 }
